Refactor exec
