+++ include/exec.h
@@ -6,7 +6,7 @@
 /*   By: reasuke <reasuke@student.42tokyo.jp>       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/09/06 12:54:48 by yliu              #+#    #+#             */
-/*   Updated: 2024/09/26 22:38:02 by reasuke          ###   ########.fr       */
+/*   Updated: 2024/10/19 16:22:19 by reasuke          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -16,6 +16,14 @@
 # include "ctx.h"
 # include "env.h"
 
-void	exec(char *input, t_ctx *ctx);
+typedef struct s_pipe_conf
+{
+	int	prev_read;
+	int	prev_write;
+	int	next_read;
+	int	next_write;
+}	t_pipe_conf;
+
+int	execute_ast_node(t_ast *node, t_ctx *ctx, t_pipe_conf *conf);
 
 #endif
