+++ src/exec/execute_ast_node.c
@@ -1,43 +1,18 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   exec.c                                             :+:      :+:    :+:   */
+/*   execute_ast_node.c                                 :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: reasuke <reasuke@student.42tokyo.jp>       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/09/06 12:54:34 by yliu              #+#    #+#             */
-/*   Updated: 2024/10/19 15:49:13 by reasuke          ###   ########.fr       */
+/*   Updated: 2024/10/19 16:25:14 by reasuke          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include <errno.h>
-#include <stdbool.h>
-#include <string.h>
+#include <stdlib.h>
 
-#include "ast.h"
-#include "ctx.h"
-#include "env.h"
 #include "exec_internal.h"
-#include "lexer.h"
-#include "parser.h"
-#include "utils.h"
-
-static bool	is_valid_last_token(t_token_list *last_token)
-{
-	if (last_token == NULL)
-		return (false);
-	if (get_token_type(last_token) == TOKEN_UNTERMINATED_QUOTE)
-	{
-		print_error("unterminated quote", get_token_value(last_token));
-		return (false);
-	}
-	if (get_token_type(last_token) == TOKEN_UNKNOWN)
-	{
-		print_error("unknown token", get_token_value(last_token));
-		return (false);
-	}
-	return (true);
-}
 
 int	execute_ast_node(t_ast *node, t_ctx *ctx, t_pipe_conf *conf)
 {
@@ -53,27 +28,4 @@ int	execute_ast_node(t_ast *node, t_ctx *ctx, t_pipe_conf *conf)
 		return (execute_subshell(node, ctx, conf));
 	else
 		return (EXIT_FAILURE);
-}
-
-void	exec(char *input, t_ctx *ctx)
-{
-	t_token_list	*token_list;
-	t_ast			*node;
-
-	token_list = lexer(input);
-	if (is_valid_last_token(ft_lstlast(token_list)) == false)
-	{
-		destroy_token_list(token_list);
-		return ;
-	}
-	node = parser(token_list);
-	if (node == NULL)
-	{
-		destroy_token_list(token_list);
-		return ;
-	}
-	if (execute_ast_node(node, ctx, NULL) == EXIT_FAILURE)
-		print_error(__func__, "failed to execute command");
-	destroy_token_list(token_list);
-	destroy_ast(node);
 }
